requirements.txt
aiohttp==3.7.4.post0
discord-py-slash-command==1.1.2
discord.py==1.7.1
dnspython==2.1.0
embed-python==0.1.12
emoji==1.2.0
Flask==1.1.2
Pillow==4.2.1
pymongo==3.11.4
requests==2.18.4
requests-file==1.4.2
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
akinator.py==4.1.0
wonderwords
termcolor
youtube_dl
opencv-python
discordlists.py